+++ src/main/java/gift/web/controller/api/MemberApiController.java
@@ -1,28 +1,18 @@
 package gift.web.controller.api;
 
-import gift.authentication.annotation.LoginMember;
 import gift.service.MemberService;
-import gift.service.WishProductService;
-import gift.web.dto.MemberDetails;
 import gift.web.dto.request.LoginRequest;
 import gift.web.dto.request.member.CreateMemberRequest;
-import gift.web.dto.request.wishproduct.UpdateWishProductRequest;
 import gift.web.dto.response.LoginResponse;
 import gift.web.dto.response.member.CreateMemberResponse;
 import gift.web.dto.response.member.ReadMemberResponse;
-import gift.web.dto.response.wishproduct.ReadAllWishProductsResponse;
-import gift.web.dto.response.wishproduct.UpdateWishProductResponse;
 import java.net.URI;
 import java.net.URISyntaxException;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -32,11 +22,9 @@ import org.springframework.web.bind.annotation.RestController;
 public class MemberApiController {
 
     private final MemberService memberService;
-    private final WishProductService wishProductService;
 
-    public MemberApiController(MemberService memberService, WishProductService wishProductService) {
+    public MemberApiController(MemberService memberService) {
         this.memberService = memberService;
-        this.wishProductService = wishProductService;
     }
 
     @PostMapping("/register")
@@ -61,24 +49,4 @@ public class MemberApiController {
         return ResponseEntity.ok(response);
     }
 
-    @GetMapping("/wishlist")
-    public ResponseEntity<ReadAllWishProductsResponse> readWishProduct(@LoginMember MemberDetails memberDetails, @PageableDefault Pageable pageable) {
-        ReadAllWishProductsResponse response = wishProductService.readAllWishProducts(memberDetails.getId(), pageable);
-        return ResponseEntity.ok(response);
-    }
-
-    @PutMapping("/wishlist/{wishProductId}")
-    public ResponseEntity<UpdateWishProductResponse> updateWishProduct(
-        @PathVariable Long wishProductId,
-        @Validated @RequestBody UpdateWishProductRequest request) {
-        UpdateWishProductResponse response = wishProductService.updateWishProduct(
-            wishProductId, request);
-        return ResponseEntity.ok(response);
-    }
-
-    @DeleteMapping("/wishlist/{wishProductId}")
-    public ResponseEntity<Void> deleteWishProduct(@PathVariable Long wishProductId) {
-        wishProductService.deleteWishProduct(wishProductId);
-        return ResponseEntity.noContent().build();
-    }
 }
